fixed bugs

+++ src/members.py
@@ -3,17 +3,26 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.exc import IntegrityError
 from flask_restful import Resource
 from flask_restful import Api
-from utils import MasonBuilder, LINK_RELATIONS_URI
-from models import *
+from utils import MasonBuilder, LINK_RELATIONS_URL, MASON
+from sqlalchemy.exc import IntegrityError, OperationalError
+from jsonschema import validate, ValidationError
 import datetime
 import enum
 import json
 
 app = Flask(__name__)
-api = Api(app)
 app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///test.db"
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+api = Api(app)
 db = SQLAlchemy(app)
+
+from models import *
+
+@event.listens_for(Engine, "connect")
+def set_sqlite_pragma(dbapi_connection, connection_record):
+    cursor = dbapi_connection.cursor()
+    cursor.execute("PRAGMA foreign_keys=ON")
+    cursor.close()
 
 class MemberBuilder(MasonBuilder):
     @staticmethod
@@ -55,7 +64,7 @@ class MemberBuilder(MasonBuilder):
     
     def add_control_edit_member(self, member):
         self.add_control("edit",
-                         f"/api/members/{member}",
+                         f"/api/members/{member}/",
                          method="PATCH",
                          encoding="json",
                          title="Edit member",
@@ -63,24 +72,25 @@ class MemberBuilder(MasonBuilder):
     
     def add_control_up_project(self, project):
         self.add_control("up",
-                         f"/api/projects/{project}/members")
+                         f"/api/projects/{project}/members/")
     
     def add_control_up_task(self, project, phase, task):
         self.add_control("up",
-                         f"/api/projects/{project}/phases/{phase}/tasks/{task}/members")
+                         f"/api/projects/{project}/phases/{phase}/tasks/{task}/members/")
     
     def add_control_delete_member(self, member, project=None, phase=None, task=None):
         if project == None:
             self.add_control("promana:delete",
-                             f"/api/members/{member}",
+                             api.url_for(MemberItem, member=member),
+                             #f"/api/members/{member}/",
                              method="DELETE")
         elif task == None:
             self.add_control("promana:delete",
-                             f"/api/projects/{project}/members/{member}",
+                             f"/api/projects/{project}/members/{member}/",
                              method="DELETE")
         else:
             self.add_control("promana:delete",
-                             f"/api/projects/{project}/phases/{phase}/tasks/{task}/members/{member}",
+                             f"/api/projects/{project}/phases/{phase}/tasks/{task}/members/{member}/",
                              method="DELETE")
 
 class MemberCollection(Resource):
@@ -92,7 +102,7 @@ class MemberCollection(Resource):
             return Response(status=501)
 
         body = MemberBuilder()
-        body.add_namespace("promana", LINK_RELATIONS_URI)
+        body.add_namespace("promana", LINK_RELATIONS_URL)
         body.add_control("self", api.url_for(MemberCollection))
         body.add_control_add_member()
         body["items"] = []
@@ -107,6 +117,16 @@ class MemberCollection(Resource):
 
     def post(self):
         # add new member
+        if not request.json:
+            return create_error_response(415, "Unsupported media type",
+                "Requests must be JSON"
+            )
+
+        try:
+            validate(request.json, MemberBuilder.member_schema())
+        except ValidationError as e:
+            return create_error_response(400, "Invalid JSON document", str(e))
+
         new_member = Members(
             name=request.json["name"])
         
@@ -114,7 +134,9 @@ class MemberCollection(Resource):
             db.session.add(new_member)
             db.session.commit()
         except IntegrityError:
-            return 409
+            return create_error_response(409, "Already exists", 
+                "Member with name '{}' already exists.".format(request.json["name"])
+            )
 
         return Response(status=201, headers={"location": api.url_for(MemberItem, member=new_member.name)})
 
@@ -124,7 +146,7 @@ class MemberItem(Resource):
         db_member = Members.query.filter_by(name=member).first()
         body = MemberBuilder(
             name=db_member.name)
-        body.add_namespace("promana", LINK_RELATIONS_URI)
+        body.add_namespace("promana", LINK_RELATIONS_URL)
         body.add_control("self", api.url_for(MemberItem, member=member))
         body.add_control("collection", api.url_for(MemberCollection))
         body.add_control_edit_member(member)
@@ -140,13 +162,20 @@ class MemberItem(Resource):
         try:
             db.session.commit()
         except IntegrityError:
-            return 409
+            return create_error_response(409, "Already exists", 
+                "Member with name '{}' already exists.".format(request.json["name"])
+            )
 
         return Response(status=204, headers={"location":api.url_for(MemberItem, member=db_member.name)})
 
     # delete member
     def delete(self, member):
         db_member = Members.query.filter_by(name=member).first()
+        if db_member is None:
+            return create_error_response(404, "Not found", 
+                "No member was found with the name {}".format(member)
+            )
+
         db.session.delete(db_member)
         db.session.commit()
         
@@ -177,3 +206,9 @@ class TaskMemberCollection(Resource):
 
 api.add_resource(MemberCollection, "/api/members/")
 api.add_resource(MemberItem, "/api/members/<member>/")
+api.add_resource(ProjectMemberCollection, "/api/projects/{project}/members/")
+api.add_resource(TaskMemberCollection, "/api/projects/{project}/phases/{phase}/tasks/{task}/members/")
+
+@app.route(LINK_RELATIONS_URL)
+def send_link_relations():
+    return "link relations"
